Remove shadow property `IFieldDefinition.Name`
This was shadowed to enforce a ReSharper hint indicating that on a field definition the Name property cannot be null. ReSharper warnings in code paths for  `IFieldDefinition` indicate the value could be null (when it cannot be); however, this condition is guarded in multiple ways on the concrete class `FieldDefinition` in Core, which would prevent any such operation in practice.